Completed Login functionality
It doesnt completely work in a dev environment but I think it is because a CORS issue, rather than try to put in a hack, ill try to set up a better dev environment later on. This commit includes several new warnings, a php Util class for getting the public variables of a class/object, an unauthorized page when trying to add, and new warnings on the login page.