Fixed deployment errors and handle 404
When I deployed to the server, the relative links were not working. I had to make some small changes in the App.js and add an htaccess file. But if I visited a page that did not exist in the app, rather than give a 404 it served a blank page instead. I had to update the htaccess again and make it more specific and handle each route individually.